requirements.txt
absl-py==2.2.2
alembic==1.15.2
amqp==5.3.1
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
Babel==2.13.1
beautifulsoup4==4.12.2
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.3
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fake-useragent==1.4.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-paginate==2024.4.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
ijson==3.3.0
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
kombu==5.5.4
libclang==18.1.1
lxml==5.3.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
playwright==1.52.0
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycocotools==2.0.8
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
redis==6.2.0
reportlab==4.4.0
requests==2.31.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.32.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
Telethon==1.40.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
zopfli==0.2.3.post1
